optimize lhs by moving video lib to delayed
